get started on deeper versioning for person detail page
