+++ templates/partials/person_list_table.html
@@ -87,7 +87,7 @@
                         {% endif %}
                         <!-- Name -->
                         <td>
-                            <a href="{% url 'view-person' object.id %}">{{ object.name.get_value }}</a>
+                            <a href="{% url 'view-person' object.uuid %}">{{ object.name.get_value }}</a>
                         </td>
                         <!-- Aliases -->
                         <td>
